refactor: move classes to panes.py
